Move common code from ble_client to esp32_ble_client base class
